pyproject.toml
[project]
name = "langchaindemo"
version = "0.1.0"
description = "LangChain框架演示项目"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "langchain>=0.1.0",
    "langchain-openai>=0.0.1",
    "langchain-community>=0.0.1",
    "python-dotenv>=1.0.0",
    "langchain-community>=0.0.1",
    "langchain-ollama>=0.0.1",
]

[tool.ruff]
line-length = 88
target-version = "py310"
